requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.1
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
redis==5.0.1

